pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "saa"
version = "2.0.0"
description = "SAA (Smart Audio Agent) - Agentic audiobook generator with Google ADK"
readme = "README.md"
requires-python = ">=3.11,<3.13"
license = {text = "MIT"}
authors = [
    {name = "Raj Sarkar"}
]
keywords = ["audiobook", "tts", "ai", "agents", "adk", "voice-cloning"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
]

dependencies = [
    "google-adk>=1.18.0",
    "replicate>=1.0.7",
    "coqui-tts>=0.24.0",
    "pydub>=0.25.1",
    "pdfplumber>=0.11.0",
    "PyPDF2>=3.0.1",
    "msgpack>=1.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "click>=8.1.0",
    "rich>=13.7.0",
    "tqdm>=4.66.0",
    "colorama>=0.4.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

api = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
]

[project.scripts]
saa = "saa.__main__:main"

[tool.setuptools]
packages = ["saa"]

[tool.setuptools.package-data]
saa = ["py.typed"]

[tool.black]
line-length = 88
target-version = ["py310", "py311"]
include = '\.pyi?$'

[tool.ruff]
line-length = 88
target-version = "py310"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
strict_optional = true

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"
addopts = "--cov=saa --cov-report=term-missing --cov-report=html --cov-fail-under=85"
